cpu-requirements.txt
torch torchvision
pillow
numpy
opencv-python
tkinter
matplotlib
scikit-image
pydicom
timm
PyWavelets
pytest-shutil 
shutils